feat: adicionado validador de Delete

+++ backend/src/Ambev.DeveloperEvaluation.Application/SaleFeatures/Commands/DeleteSale/DeleteSaleCommandHandler.cs
@@ -1,5 +1,6 @@
 using Ambev.DeveloperEvaluation.Domain.Models.SaleAggregate.Events;
 using Ambev.DeveloperEvaluation.Domain.Models.SaleAggregate.Repositories;
+using FluentValidation;
 using MediatR;
 
 namespace Ambev.DeveloperEvaluation.Application.SaleFeatures.Commands.DeleteSale;
@@ -9,14 +10,20 @@ public class DeleteSaleCommandHandler(ISaleRepository saleRepository, IMediator 
 {
     public async Task<bool> Handle(DeleteSaleCommand command, CancellationToken cancellationToken)
     {
+        var validator = new DeleteSaleValidator();
+        var validationResult = await validator.ValidateAsync(command, cancellationToken);
+
+        if (!validationResult.IsValid)
+            throw new ValidationException(validationResult.Errors);
+
         var sale = await saleRepository.GetByIdAsync(command.Id, cancellationToken);
         if (sale == null) throw new Exception($"Sale with Id {command.Id} not found.");
 
         var deletedSale = await saleRepository.DeleteAsync(sale.Id, cancellationToken);
         if (!deletedSale) throw new Exception($"Sale with Id {command.Id} not deleted.");
-        
+
         await mediator.Publish(new SaleDeletedEvent(sale.SaleNumber), cancellationToken);
-        
+
         return true;
     }
 }
